Miner ai redone
Fixes lots of issues regarding hiring, firing, if you didnt notice any issues before, then you should barely notice this change
