+++ src/main/java/com/resimulators/simukraft/common/entity/goals/BakerGoal.java
@@ -3,9 +3,9 @@ package com.resimulators.simukraft.common.entity.goals;
 import com.resimulators.simukraft.common.entity.sim.SimEntity;
 import com.resimulators.simukraft.common.enums.Seed;
 import com.resimulators.simukraft.common.jobs.JobBaker;
-import com.resimulators.simukraft.common.jobs.JobFarmer;
 import com.resimulators.simukraft.common.jobs.Profession;
 import com.resimulators.simukraft.common.jobs.core.Activity;
+import com.resimulators.simukraft.common.jobs.reworked.JobFarmer;
 import com.resimulators.simukraft.common.tileentity.TileFarmer;
 import com.resimulators.simukraft.common.world.Faction;
 import com.resimulators.simukraft.common.world.SavedWorldData;
