+++ src/main/java/com/resimulators/simukraft/common/entity/sim/WorkingController.java
@@ -44,6 +44,8 @@ public class WorkingController {
                     if (!job.nightShift()) {
                         sim.setActivity(Activity.FORCE_STOP);
                     }
+                } else if (!job.nightShift()){
+                    sim.setActivity(Activity.WORKING);
                 }
             }
         }
