+++ src/main/java/com/resimulators/simukraft/init/ModJobs.java
@@ -5,6 +5,7 @@ import com.resimulators.simukraft.common.entity.sim.SimEntity;
 import com.resimulators.simukraft.common.jobs.Profession;
 import com.resimulators.simukraft.common.jobs.core.IReworkedJob;
 import com.resimulators.simukraft.common.jobs.reworked.JobFarmer;
+import com.resimulators.simukraft.common.jobs.reworked.JobMiner;
 
 import java.util.Map;
 import java.util.function.Function;
@@ -12,6 +13,7 @@ import java.util.function.Function;
 public class ModJobs {
     public static final Map<Integer, Function<SimEntity, IReworkedJob>> JOB_LOOKUP = new ImmutableMap.Builder<Integer, Function<SimEntity, IReworkedJob>>()
             .put(Profession.FARMER.getId(), JobFarmer::new)
+            .put(Profession.MINER.getId(), JobMiner::new)
             .build();
 
 }
